Pass nullptr to glTexImage2D instead
For the null data case, I believe the correct thing to do is also just pass nullptr to glTexImage2D. Passing null is valid and common (when you just want to allocate a frame buffer, for example). I believe that creating and destroy a temporary buffer is actively wrong and actually unsafe, because the driver may try to access this data asynchronously. In any case, it's unnecessary.